Changes behaviour of interface implementations
Previously, you could register either an instance or a class name as an
implementation of an interface; either way would result in a
singleton instance being registered.

Now you can only register a class name and a new instance of this will
be resolved each time. Singleton behaviour can still be achieved by
explicitly resolving one instance and passing it to 'registerSingleton'.